rename the collector objects to factories
