+++ src/main/java/com/blackducksoftware/integration/hub/notification/content/collector/VulnerabilityDetailFactory.java
@@ -25,27 +25,21 @@ package com.blackducksoftware.integration.hub.notification.content.collector;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import com.blackducksoftware.integration.hub.notification.content.NotificationContent;
 import com.blackducksoftware.integration.hub.notification.content.NotificationContentDetail;
-import com.blackducksoftware.integration.hub.notification.content.RuleViolationClearedNotificationContent;
+import com.blackducksoftware.integration.hub.notification.content.VulnerabilityNotificationContent;
 
-public class RuleViolationClearedDetailCollector extends NotificationDetailCollector {
+public class VulnerabilityDetailFactory extends NotificationDetailFactory {
 
     @Override
     public List<NotificationContentDetail> createDetails(final NotificationContent notificationContent) {
-        final RuleViolationClearedNotificationContent content = (RuleViolationClearedNotificationContent) notificationContent;
-        final Map<String, String> uriToName = content.policyInfos.stream().collect(Collectors.toMap(policyInfo -> policyInfo.policy, policyInfo -> policyInfo.policyName));
+        final VulnerabilityNotificationContent content = (VulnerabilityNotificationContent) notificationContent;
         final List<NotificationContentDetail> details = new ArrayList<>();
-        content.componentVersionStatuses.forEach(componentVersionStatus -> {
-            componentVersionStatus.policies.forEach(policyUri -> {
-                final String policyName = uriToName.get(policyUri);
-                details.add(NotificationContentDetail.createDetail(content, content.projectName, content.projectVersionName, content.projectVersion, componentVersionStatus.componentName,
-                        componentVersionStatus.component, componentVersionStatus.componentVersionName, componentVersionStatus.componentVersion, policyName,
-                        policyUri, null, componentVersionStatus.componentIssueLink, null));
-            });
+        content.affectedProjectVersions.forEach(projectVersion -> {
+            details.add(NotificationContentDetail.createDetail(content, projectVersion.projectName, projectVersion.projectVersionName, projectVersion.projectVersion, content.componentName,
+                    null, content.versionName, content.componentVersion, null, null,
+                    content.componentVersionOriginName, projectVersion.componentIssueUrl, content.componentVersionOriginId));
         });
         return details;
     }
